[gameplay] fix bug where a dead zombie would still hurt the player on collision
After dying, zombie would try to respawn, but sometimes all the respawn slots
would be taken by other entities, and the respawn code doesn't attempt to
resolve such situations and simply waits and retries. However, the zombie would
stay in place until respawned, and would still hurt the player on collision.
This fixes the "hurt" part, but doesn't hide the zombie texture, because
a proper fix would be to make a dead state for zombies instead.

@TODO: dead texture for zombies